Move label selector function to separate file
